fix(series): correct page response dto totalteaching type

+++ server/src/main/java/org/eni/koinoniadaily/modules/series/dto/SeriesPageResponse.java
@@ -23,5 +23,5 @@ public class SeriesPageResponse {
 
   private LocalDateTime updatedAt;
 
-  private Long totalTeachings;
+  private Integer totalTeachings;
 }
